+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Upcoming...
 
 - ... <!-- Add new lines here. Version number will be decided later -->
+- feat: Add new `types` module including `DeepPartial`, `SparseArray`,
+  `ReadonlySparseArray`, `SaneRecord`, `ExcludesFalse`, and more
 
 ## 4.8.2